+++ chat/kit/src/main/java/cn/wildfire/chat/kit/conversation/ConversationActivity.java
@@ -1,240 +1,29 @@
 package cn.wildfire.chat.kit.conversation;
 
-import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
-import android.os.Bundle;
-import android.os.Handler;
-import android.text.SpannableString;
-import android.text.Spanned;
-import android.text.TextUtils;
 import android.view.MenuItem;
-import android.view.MotionEvent;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.LinearLayout;
-import android.widget.TextView;
-import android.widget.Toast;
 
-import androidx.annotation.Nullable;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProviders;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
-import androidx.recyclerview.widget.SimpleItemAnimator;
-import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
-import com.afollestad.materialdialogs.MaterialDialog;
-
-import java.util.List;
-import java.util.Map;
-
-import butterknife.BindView;
-import butterknife.OnTouch;
-import cn.wildfire.chat.kit.ChatManagerHolder;
 import cn.wildfire.chat.kit.IMServiceStatusViewModel;
 import cn.wildfire.chat.kit.WfcBaseActivity;
-import cn.wildfire.chat.kit.WfcUIKit;
-import cn.wildfire.chat.kit.channel.ChannelViewModel;
-import cn.wildfire.chat.kit.chatroom.ChatRoomViewModel;
-import cn.wildfire.chat.kit.common.OperateResult;
-import cn.wildfire.chat.kit.conversation.ext.core.ConversationExtension;
-import cn.wildfire.chat.kit.conversation.mention.MentionSpan;
-import cn.wildfire.chat.kit.conversation.message.model.UiMessage;
-import cn.wildfire.chat.kit.conversation.multimsg.MultiMessageAction;
-import cn.wildfire.chat.kit.conversation.multimsg.MultiMessageActionManager;
-import cn.wildfire.chat.kit.group.GroupViewModel;
-import cn.wildfire.chat.kit.third.utils.UIUtils;
-import cn.wildfire.chat.kit.user.UserInfoActivity;
-import cn.wildfire.chat.kit.user.UserViewModel;
-import cn.wildfire.chat.kit.viewmodel.MessageViewModel;
-import cn.wildfire.chat.kit.viewmodel.SettingViewModel;
-import cn.wildfire.chat.kit.widget.InputAwareLayout;
-import cn.wildfire.chat.kit.widget.KeyboardAwareLinearLayout;
 import cn.wildfirechat.chat.R;
-import cn.wildfirechat.message.MessageContent;
-import cn.wildfirechat.message.TypingMessageContent;
-import cn.wildfirechat.message.core.MessageDirection;
-import cn.wildfirechat.message.core.PersistFlag;
-import cn.wildfirechat.message.notification.TipNotificationContent;
-import cn.wildfirechat.model.ChannelInfo;
-import cn.wildfirechat.model.ChatRoomInfo;
 import cn.wildfirechat.model.Conversation;
-import cn.wildfirechat.model.GroupInfo;
-import cn.wildfirechat.model.GroupMember;
-import cn.wildfirechat.model.UserInfo;
 import cn.wildfirechat.remote.ChatManager;
-import cn.wildfirechat.remote.UserSettingScope;
 
-public class ConversationActivity extends WfcBaseActivity implements
-        KeyboardAwareLinearLayout.OnKeyboardShownListener,
-        KeyboardAwareLinearLayout.OnKeyboardHiddenListener,
-        ConversationMessageAdapter.OnPortraitClickListener,
-        ConversationMessageAdapter.OnPortraitLongClickListener,
-        ConversationInputPanel.OnConversationInputPanelStateChangeListener,
-        ConversationMessageAdapter.OnMessageCheckListener {
-
-    public static final int REQUEST_PICK_MENTION_CONTACT = 100;
-
-    private Conversation conversation;
-    private boolean loadingNewMessage;
-    private boolean shouldContinueLoadNewMessage = false;
-
-    private static final int MESSAGE_LOAD_COUNT_PER_TIME = 20;
-    private static final int MESSAGE_LOAD_AROUND = 10;
-
-    @BindView(R.id.rootLinearLayout)
-    InputAwareLayout rootLinearLayout;
-    @BindView(R.id.swipeRefreshLayout)
-    SwipeRefreshLayout swipeRefreshLayout;
-    @BindView(R.id.msgRecyclerView)
-    RecyclerView recyclerView;
-
-    @BindView(R.id.inputPanelFrameLayout)
-    ConversationInputPanel inputPanel;
-
-    @BindView(R.id.multiMessageActionContainerLinearLayout)
-    LinearLayout multiMessageActionContainerLinearLayout;
-
-    private ConversationMessageAdapter adapter;
-    private boolean moveToBottom = true;
-    private ConversationViewModel conversationViewModel;
-    private SettingViewModel settingViewModel;
-    private MessageViewModel messageViewModel;
-    private UserViewModel userViewModel;
+public class ConversationActivity extends WfcBaseActivity {
     private boolean isInitialized = false;
-    private ChatRoomViewModel chatRoomViewModel;
-
-    private Handler handler;
-    private long initialFocusedMessageId;
-    // 用户channel主发起，针对某个用户的会话
-    private String channelPrivateChatUser;
-    private String conversationTitle = "";
-    private SharedPreferences sharedPreferences;
-    private LinearLayoutManager layoutManager;
-
-    private GroupInfo groupInfo;
-    private boolean showGroupMemberName = false;
-
-    private Observer<UiMessage> messageLiveDataObserver = new Observer<UiMessage>() {
-        @Override
-        public void onChanged(@Nullable UiMessage uiMessage) {
-            if (!isMessageInCurrentConversation(uiMessage)) {
-                return;
-            }
-            MessageContent content = uiMessage.message.content;
-            if (isDisplayableMessage(uiMessage)) {
-                // 消息定位时，如果收到新消息、或者发送消息，需要重新加载消息列表
-                if (shouldContinueLoadNewMessage) {
-                    shouldContinueLoadNewMessage = false;
-                    reloadMessage();
-                    return;
-                }
-                adapter.addNewMessage(uiMessage);
-                if (moveToBottom || uiMessage.message.sender.equals(ChatManager.Instance().getUserId())) {
-                    UIUtils.postTaskDelay(() -> {
-
-                                int position = adapter.getItemCount() - 1;
-                                if (position < 0) {
-                                    return;
-                                }
-                                recyclerView.scrollToPosition(position);
-                            },
-                            100);
-                }
-            }
-            if (content instanceof TypingMessageContent && uiMessage.message.direction == MessageDirection.Receive) {
-                updateTypingStatusTitle((TypingMessageContent) content);
-            } else {
-                resetConversationTitle();
-            }
-
-            if (uiMessage.message.direction == MessageDirection.Receive) {
-                conversationViewModel.clearUnreadStatus(conversation);
-            }
-        }
-    };
-    private Observer<UiMessage> messageUpdateLiveDatObserver = new Observer<UiMessage>() {
-        @Override
-        public void onChanged(@Nullable UiMessage uiMessage) {
-            if (!isMessageInCurrentConversation(uiMessage)) {
-                return;
-            }
-            if (isDisplayableMessage(uiMessage)) {
-                adapter.updateMessage(uiMessage);
-            }
-        }
-    };
-
-    private Observer<UiMessage> messageRemovedLiveDataObserver = new Observer<UiMessage>() {
-        @Override
-        public void onChanged(@Nullable UiMessage uiMessage) {
-            if (!isMessageInCurrentConversation(uiMessage)) {
-                return;
-            }
-            if (isDisplayableMessage(uiMessage)) {
-                adapter.removeMessage(uiMessage);
-            }
-        }
-    };
-
-    private boolean isDisplayableMessage(UiMessage uiMessage) {
-        MessageContent content = uiMessage.message.content;
-        if (content.getPersistFlag() == PersistFlag.Persist
-                || content.getPersistFlag() == PersistFlag.Persist_And_Count) {
-            return true;
-        }
-        return false;
-    }
-
-    private Observer<Map<String, String>> mediaUploadedLiveDataObserver = new Observer<Map<String, String>>() {
-        @Override
-        public void onChanged(@Nullable Map<String, String> stringStringMap) {
-            for (Map.Entry<String, String> entry : stringStringMap.entrySet()) {
-                sharedPreferences.edit()
-                        .putString(entry.getKey(), entry.getValue())
-                        .apply();
-            }
-
-        }
-    };
-
-    private Observer<Conversation> clearConversationMessageObserver = new Observer<Conversation>() {
-        @Override
-        public void onChanged(Conversation conversation) {
-            if (conversation.equals(ConversationActivity.this.conversation)) {
-                adapter.setMessages(null);
-                adapter.notifyDataSetChanged();
-            }
-        }
-    };
-
-    private Observer<List<UserInfo>> userInfoUpdateLiveDataObserver = new Observer<List<UserInfo>>() {
-        @Override
-        public void onChanged(@Nullable List<UserInfo> userInfos) {
-            if (conversation.type == Conversation.ConversationType.Single) {
-                conversationTitle = null;
-                setTitle();
-            }
-            int start = layoutManager.findFirstVisibleItemPosition();
-            int end = layoutManager.findLastVisibleItemPosition();
-            adapter.notifyItemRangeChanged(start, end - start + 1, userInfos);
-        }
-    };
-
-    private boolean isMessageInCurrentConversation(UiMessage message) {
-        if (conversation == null || message == null || message.message == null) {
-            return false;
-        }
-        return conversation.equals(message.message.conversation);
-    }
+    private ConversationFragment conversationFragment;
+    private Conversation conversation;
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        // you can setup background here
+    protected int contentLayout() {
+        return R.layout.fragment_container_activity;
+    }
+
+    private void setConversationBackground() {
+        // you can setup your conversation background here
 //        getWindow().setBackgroundDrawableResource(R.mipmap.splash);
     }
 
@@ -247,25 +36,64 @@ public class ConversationActivity extends WfcBaseActivity implements
                 isInitialized = true;
             }
         });
-    }
+        conversationFragment = new ConversationFragment();
+        getSupportFragmentManager().beginTransaction()
+                .add(R.id.containerFrameLayout, conversationFragment, "content")
+                .commit();
 
-    private void init() {
-        initView();
-        sharedPreferences = getSharedPreferences("sticker", Context.MODE_PRIVATE);
-        Intent intent = getIntent();
-        conversation = intent.getParcelableExtra("conversation");
-        conversationTitle = intent.getStringExtra("conversationTitle");
-        initialFocusedMessageId = intent.getLongExtra("toFocusMessageId", -1);
-        if (conversation == null) {
-            finish();
-        }
-        setupConversation(conversation);
-        conversationViewModel.clearUnreadStatus(conversation);
+        setConversationBackground();
     }
 
     @Override
-    protected int contentLayout() {
-        return R.layout.conversation_activity;
+    protected int menu() {
+        return R.menu.conversation;
+    }
+
+    public ConversationFragment getConversationFragment() {
+        return conversationFragment;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        if (item.getItemId() == R.id.menu_conversation_info) {
+            showConversationInfo();
+            return true;
+        }
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
+    public void onBackPressed() {
+        if (!conversationFragment.onBackPressed()) {
+            super.onBackPressed();
+        }
+    }
+
+    private void showConversationInfo() {
+        Intent intent = new Intent(this, ConversationInfoActivity.class);
+        intent.putExtra("conversationInfo", ChatManager.Instance().getConversation(conversation));
+        startActivity(intent);
+    }
+
+    @Override
+    protected void onNewIntent(Intent intent) {
+        super.onNewIntent(intent);
+        conversation = intent.getParcelableExtra("conversation");
+        long initialFocusedMessageId = intent.getLongExtra("toFocusMessageId", -1);
+        String channelPrivateChatUser = intent.getStringExtra("channelPrivateChatUser");
+        conversationFragment.setupConversation(conversation, null, initialFocusedMessageId, channelPrivateChatUser);
+    }
+
+
+    private void init() {
+        Intent intent = getIntent();
+        conversation = intent.getParcelableExtra("conversation");
+        String conversationTitle = intent.getStringExtra("conversationTitle");
+        long initialFocusedMessageId = intent.getLongExtra("toFocusMessageId", -1);
+        if (conversation == null) {
+            finish();
+        }
+        conversationFragment.setupConversation(conversation, conversationTitle, initialFocusedMessageId, null);
     }
 
     public static Intent buildConversationIntent(Context context, Conversation.ConversationType type, String target, int line) {
@@ -283,531 +111,10 @@ public class ConversationActivity extends WfcBaseActivity implements
     }
 
     public static Intent buildConversationIntent(Context context, Conversation conversation, String channelPrivateChatUser, long toFocusMessageId) {
-        Intent intent = new Intent(context, ConversationActivity.class);
+        Intent intent = new Intent(context, ConversationFragment.class);
         intent.putExtra("conversation", conversation);
         intent.putExtra("toFocusMessageId", toFocusMessageId);
         intent.putExtra("channelPrivateChatUser", channelPrivateChatUser);
         return intent;
-    }
-
-    @Override
-    protected void onNewIntent(Intent intent) {
-        super.onNewIntent(intent);
-        conversation = intent.getParcelableExtra("conversation");
-        initialFocusedMessageId = intent.getLongExtra("toFocusMessageId", -1);
-        channelPrivateChatUser = intent.getStringExtra("channelPrivateChatUser");
-        setupConversation(conversation);
-    }
-
-    private void initView() {
-        handler = new Handler();
-        rootLinearLayout.addOnKeyboardShownListener(this);
-
-        swipeRefreshLayout.setOnRefreshListener(() -> {
-            if (adapter.getMessages() == null || adapter.getMessages().isEmpty()) {
-                swipeRefreshLayout.setRefreshing(false);
-                return;
-            }
-            loadMoreOldMessages();
-        });
-
-        // message list
-        adapter = new ConversationMessageAdapter(this);
-        adapter.setOnPortraitClickListener(this);
-        adapter.setOnPortraitLongClickListener(this);
-        adapter.setOnMessageCheckListener(this);
-        layoutManager = new LinearLayoutManager(this);
-        recyclerView.setLayoutManager(layoutManager);
-        recyclerView.setAdapter(adapter);
-        ((SimpleItemAnimator) recyclerView.getItemAnimator()).setSupportsChangeAnimations(false);
-        recyclerView.addOnScrollListener(new RecyclerView.OnScrollListener() {
-            @Override
-            public void onScrollStateChanged(RecyclerView recyclerView, int newState) {
-                super.onScrollStateChanged(recyclerView, newState);
-                // 向上滑动，不在底部，收到消息时，不滑动到底部, 发送消息时，可以强制置为true
-                if (newState != RecyclerView.SCROLL_STATE_IDLE) {
-                    return;
-                }
-                if (!recyclerView.canScrollVertically(1)) {
-                    moveToBottom = true;
-                    if (initialFocusedMessageId != -1 && !loadingNewMessage && shouldContinueLoadNewMessage) {
-                        int lastVisibleItem = layoutManager.findLastCompletelyVisibleItemPosition();
-                        if (lastVisibleItem > adapter.getItemCount() - 3) {
-                            loadMoreNewMessages();
-                        }
-                    }
-                } else {
-                    moveToBottom = false;
-                }
-            }
-        });
-
-        inputPanel.init(this, rootLinearLayout);
-        inputPanel.setOnConversationInputPanelStateChangeListener(this);
-    }
-
-    private void setupConversation(Conversation conversation) {
-        if (conversationViewModel == null) {
-            settingViewModel = ViewModelProviders.of(this).get(SettingViewModel.class);
-            conversationViewModel = WfcUIKit.getAppScopeViewModel(ConversationViewModel.class);
-            conversationViewModel.clearConversationMessageLiveData().observeForever(clearConversationMessageObserver);
-            messageViewModel = ViewModelProviders.of(this).get(MessageViewModel.class);
-
-            messageViewModel.messageLiveData().observeForever(messageLiveDataObserver);
-            messageViewModel.messageUpdateLiveData().observeForever(messageUpdateLiveDatObserver);
-            messageViewModel.messageRemovedLiveData().observeForever(messageRemovedLiveDataObserver);
-            messageViewModel.mediaUpdateLiveData().observeForever(mediaUploadedLiveDataObserver);
-
-            userViewModel = ViewModelProviders.of(this).get(UserViewModel.class);
-            userViewModel.userInfoLiveData().observeForever(userInfoUpdateLiveDataObserver);
-        }
-
-        if (conversation.type == Conversation.ConversationType.Group) {
-            GroupViewModel groupViewModel = ViewModelProviders.of(this).get(GroupViewModel.class);
-            groupInfo = groupViewModel.getGroupInfo(conversation.target, false);
-            groupViewModel.groupInfoUpdateLiveData().observe(this, groupInfos -> {
-                for (GroupInfo info : groupInfos) {
-                    if (info.target.equals(groupInfo.target)) {
-                        groupInfo = info;
-                        if (groupInfo.mute == 1) {
-                            GroupMember groupMember = groupViewModel.getGroupMember(groupInfo.target, userViewModel.getUserId());
-                            if (groupMember.type != GroupMember.GroupMemberType.Owner && groupMember.type != GroupMember.GroupMemberType.Manager) {
-                                inputPanel.disableInput("全员禁言中");
-                            } else {
-                                inputPanel.enableInput();
-                            }
-                        } else {
-                            inputPanel.enableInput();
-                        }
-                        setTitle();
-                        adapter.notifyDataSetChanged();
-                    }
-                }
-            });
-
-            showGroupMemberName = "1".equals(userViewModel.getUserSetting(UserSettingScope.GroupHideNickname, groupInfo.target));
-            settingViewModel.settingUpdatedLiveData().observe(this, o -> {
-                boolean showGroupMemberName = "1".equals(userViewModel.getUserSetting(UserSettingScope.GroupHideNickname, groupInfo.target));
-                if (this.showGroupMemberName != showGroupMemberName) {
-                    this.showGroupMemberName = showGroupMemberName;
-                    adapter.notifyDataSetChanged();
-                }
-            });
-
-            if (groupInfo.mute == 1) {
-                GroupMember groupMember = groupViewModel.getGroupMember(groupInfo.target, userViewModel.getUserId());
-                if (groupMember.type != GroupMember.GroupMemberType.Owner && groupMember.type != GroupMember.GroupMemberType.Manager) {
-                    inputPanel.disableInput("全员禁言中");
-                }
-            }
-        }
-
-        inputPanel.setupConversation(conversation);
-
-        MutableLiveData<List<UiMessage>> messages;
-        if (initialFocusedMessageId != -1) {
-            shouldContinueLoadNewMessage = true;
-            messages = conversationViewModel.loadAroundMessages(conversation, channelPrivateChatUser, initialFocusedMessageId, MESSAGE_LOAD_AROUND);
-        } else {
-            messages = conversationViewModel.getMessages(conversation, channelPrivateChatUser);
-        }
-
-        // load message
-        swipeRefreshLayout.setRefreshing(true);
-        messages.observe(this, uiMessages -> {
-            swipeRefreshLayout.setRefreshing(false);
-            adapter.setMessages(uiMessages);
-            adapter.notifyDataSetChanged();
-
-            if (adapter.getItemCount() > 1) {
-                int initialMessagePosition;
-                if (initialFocusedMessageId != -1) {
-                    initialMessagePosition = adapter.getMessagePosition(initialFocusedMessageId);
-                    if (initialMessagePosition != -1) {
-                        recyclerView.scrollToPosition(initialMessagePosition);
-                        adapter.highlightFocusMessage(initialMessagePosition);
-                    }
-                } else {
-                    moveToBottom = true;
-                    recyclerView.scrollToPosition(adapter.getItemCount() - 1);
-                }
-            }
-        });
-        if (conversation.type == Conversation.ConversationType.ChatRoom) {
-            joinChatRoom();
-        }
-
-        setTitle();
-    }
-
-    private void joinChatRoom() {
-        chatRoomViewModel = ViewModelProviders.of(this).get(ChatRoomViewModel.class);
-        chatRoomViewModel.joinChatRoom(conversation.target)
-                .observe(this, new Observer<OperateResult<Boolean>>() {
-                    @Override
-                    public void onChanged(@Nullable OperateResult<Boolean> booleanOperateResult) {
-                        if (booleanOperateResult.isSuccess()) {
-                            String welcome = "欢迎 %s 加入聊天室";
-                            TipNotificationContent content = new TipNotificationContent();
-                            String userId = userViewModel.getUserId();
-                            UserInfo userInfo = userViewModel.getUserInfo(userId, false);
-                            if (userInfo != null) {
-                                content.tip = String.format(welcome, userInfo.displayName);
-                            } else {
-                                content.tip = String.format(welcome, "<" + userId + ">");
-                            }
-                            messageViewModel.sendMessage(conversation, content);
-                            loadMoreOldMessages();
-                            setChatRoomConversationTitle();
-
-                        } else {
-                            Toast.makeText(ConversationActivity.this, "加入聊天室失败", Toast.LENGTH_SHORT).show();
-                            finish();
-                        }
-                    }
-                });
-    }
-
-    private void quitChatRoom() {
-        String welcome = "%s 离开了聊天室";
-        TipNotificationContent content = new TipNotificationContent();
-        String userId = userViewModel.getUserId();
-        UserInfo userInfo = userViewModel.getUserInfo(userId, false);
-        if (userInfo != null) {
-            content.tip = String.format(welcome, userInfo.displayName);
-        } else {
-            content.tip = String.format(welcome, "<" + userId + ">");
-        }
-        messageViewModel.sendMessage(conversation, content);
-        chatRoomViewModel.quitChatRoom(conversation.target);
-    }
-
-    private void setChatRoomConversationTitle() {
-        chatRoomViewModel.getChatRoomInfo(conversation.target, System.currentTimeMillis())
-                .observe(this, chatRoomInfoOperateResult -> {
-                    if (chatRoomInfoOperateResult.isSuccess()) {
-                        ChatRoomInfo chatRoomInfo = chatRoomInfoOperateResult.getResult();
-                        conversationTitle = chatRoomInfo.title;
-                        setTitle(conversationTitle);
-                    }
-                });
-    }
-
-    private void setTitle() {
-        if (!TextUtils.isEmpty(conversationTitle)) {
-            setTitle(conversationTitle);
-        }
-
-        if (conversation.type == Conversation.ConversationType.Single) {
-            UserInfo userInfo = ChatManagerHolder.gChatManager.getUserInfo(conversation.target, false);
-            conversationTitle = userViewModel.getUserDisplayName(userInfo);
-        } else if (conversation.type == Conversation.ConversationType.Group) {
-            if (groupInfo != null) {
-                conversationTitle = groupInfo.name + "(" + groupInfo.memberCount + "人)";
-            }
-        } else if (conversation.type == Conversation.ConversationType.Channel) {
-            ChannelViewModel channelViewModel = ViewModelProviders.of(this).get(ChannelViewModel.class);
-            ChannelInfo channelInfo = channelViewModel.getChannelInfo(conversation.target, false);
-            if (channelInfo != null) {
-                conversationTitle = channelInfo.name;
-            }
-
-            if (!TextUtils.isEmpty(channelPrivateChatUser)) {
-                UserInfo channelPrivateChatUserInfo = userViewModel.getUserInfo(channelPrivateChatUser, false);
-                if (channelPrivateChatUserInfo != null) {
-                    conversationTitle += "@" + userViewModel.getUserDisplayName(channelPrivateChatUserInfo);
-                } else {
-                    conversationTitle += "@<" + channelPrivateChatUser + ">";
-                }
-            }
-        }
-        setTitle(conversationTitle);
-    }
-
-    @Override
-    protected int menu() {
-        return R.menu.conversation;
-    }
-
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        if (item.getItemId() == R.id.menu_conversation_info) {
-            showConversationInfo();
-            return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
-
-    private void showConversationInfo() {
-        Intent intent = new Intent(ConversationActivity.this, ConversationInfoActivity.class);
-        intent.putExtra("conversationInfo", ChatManager.Instance().getConversation(conversation));
-        startActivity(intent);
-    }
-
-    @OnTouch({R.id.contentLayout, R.id.msgRecyclerView})
-    boolean onTouch(View view, MotionEvent event) {
-        if (event.getAction() == MotionEvent.ACTION_DOWN && inputPanel.extension.canHideOnScroll()) {
-            inputPanel.collapse();
-        }
-        return false;
-    }
-
-    @Override
-    public void onPortraitClick(UserInfo userInfo) {
-        Intent intent = new Intent(this, UserInfoActivity.class);
-        intent.putExtra("userInfo", userInfo);
-        startActivity(intent);
-    }
-
-    @Override
-    public void onPortraitLongClick(UserInfo userInfo) {
-        // TODO panel insert
-        int position = inputPanel.editText.getSelectionEnd();
-        position = position >= 0 ? position : 0;
-        if (conversation.type == Conversation.ConversationType.Group) {
-            SpannableString spannableString = mentionSpannable(userInfo);
-            inputPanel.editText.getEditableText().insert(position, spannableString);
-        } else {
-            inputPanel.editText.getEditableText().insert(position, userViewModel.getUserDisplayName(userInfo));
-        }
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode >= ConversationExtension.REQUEST_CODE_MIN) {
-            inputPanel.extension.onActivityResult(requestCode, resultCode, data);
-            return;
-        } else if (resultCode == Activity.RESULT_OK && requestCode == REQUEST_PICK_MENTION_CONTACT) {
-            boolean isMentionAll = data.getBooleanExtra("mentionAll", false);
-            SpannableString spannableString;
-            if (isMentionAll) {
-                spannableString = mentionAllSpannable();
-            } else {
-                String userId = data.getStringExtra("userId");
-                UserInfo userInfo = userViewModel.getUserInfo(userId, false);
-                spannableString = mentionSpannable(userInfo);
-            }
-            int position = inputPanel.editText.getSelectionEnd();
-            position = position > 0 ? position - 1 : 0;
-            inputPanel.editText.getEditableText().replace(position, position + 1, spannableString);
-            return;
-        }
-        super.onActivityResult(requestCode, resultCode, data);
-    }
-
-    private SpannableString mentionAllSpannable() {
-        String text = "@所有人 ";
-        SpannableString spannableString = new SpannableString(text);
-        spannableString.setSpan(new MentionSpan(true), 0, spannableString.length(), Spanned.SPAN_INCLUSIVE_EXCLUSIVE);
-        return spannableString;
-    }
-
-    private SpannableString mentionSpannable(UserInfo userInfo) {
-        String text = "@" + userInfo.displayName + " ";
-        SpannableString spannableString = new SpannableString(text);
-        spannableString.setSpan(new MentionSpan(userInfo.uid), 0, spannableString.length(), Spanned.SPAN_INCLUSIVE_EXCLUSIVE);
-        return spannableString;
-    }
-
-    @Override
-    protected void onPause() {
-        super.onPause();
-        inputPanel.onActivityPause();
-        messageViewModel.stopPlayAudio();
-    }
-
-    @Override
-    protected void onDestroy() {
-        if (conversation.type == Conversation.ConversationType.ChatRoom) {
-            quitChatRoom();
-        }
-
-        super.onDestroy();
-        messageViewModel.messageLiveData().removeObserver(messageLiveDataObserver);
-        messageViewModel.messageUpdateLiveData().removeObserver(messageUpdateLiveDatObserver);
-        messageViewModel.messageRemovedLiveData().removeObserver(messageRemovedLiveDataObserver);
-        messageViewModel.mediaUpdateLiveData().removeObserver(mediaUploadedLiveDataObserver);
-        userViewModel.userInfoLiveData().removeObserver(userInfoUpdateLiveDataObserver);
-        conversationViewModel.clearConversationMessageLiveData().removeObserver(clearConversationMessageObserver);
-        inputPanel.onDestroy();
-    }
-
-    @Override
-    public void onBackPressed() {
-
-        if (rootLinearLayout.getCurrentInput() != null) {
-            rootLinearLayout.hideAttachedInput(true);
-            inputPanel.collapse();
-        } else if (multiMessageActionContainerLinearLayout.getVisibility() == View.VISIBLE) {
-            toggleConversationMode();
-        } else {
-            super.onBackPressed();
-        }
-    }
-
-    @Override
-    public void onKeyboardShown() {
-        inputPanel.onKeyboardShown();
-        recyclerView.scrollToPosition(adapter.getItemCount() - 1);
-    }
-
-    @Override
-    public void onKeyboardHidden() {
-        inputPanel.onKeyboardHidden();
-    }
-
-    private void reloadMessage() {
-        conversationViewModel.getMessages(conversation, channelPrivateChatUser).observe(this, uiMessages -> {
-            adapter.setMessages(uiMessages);
-            adapter.notifyDataSetChanged();
-        });
-    }
-
-    private void loadMoreOldMessages() {
-        long fromMessageId = Long.MAX_VALUE;
-        long fromMessageUid = Long.MAX_VALUE;
-        if (adapter.getMessages() != null && !adapter.getMessages().isEmpty()) {
-            fromMessageId = adapter.getItem(0).message.messageId;
-            fromMessageUid = adapter.getItem(0).message.messageUid;
-        }
-
-        conversationViewModel.loadOldMessages(conversation, channelPrivateChatUser, fromMessageId, fromMessageUid, MESSAGE_LOAD_COUNT_PER_TIME)
-                .observe(this, uiMessages -> {
-                    adapter.addMessagesAtHead(uiMessages);
-
-                    swipeRefreshLayout.setRefreshing(false);
-                });
-    }
-
-    private void loadMoreNewMessages() {
-        loadingNewMessage = true;
-        adapter.showLoadingNewMessageProgressBar();
-        conversationViewModel.loadNewMessages(conversation, channelPrivateChatUser, adapter.getItem(adapter.getItemCount() - 2).message.messageId, MESSAGE_LOAD_COUNT_PER_TIME)
-                .observe(this, messages -> {
-                    loadingNewMessage = false;
-                    adapter.dismissLoadingNewMessageProgressBar();
-
-                    if (messages == null || messages.isEmpty()) {
-                        shouldContinueLoadNewMessage = false;
-                    }
-                    if (messages != null && !messages.isEmpty()) {
-                        adapter.addMessagesAtTail(messages);
-                    }
-                });
-    }
-
-    private void updateTypingStatusTitle(TypingMessageContent typingMessageContent) {
-        String typingDesc = "";
-        switch (typingMessageContent.getType()) {
-            case TypingMessageContent.TYPING_TEXT:
-                typingDesc = "对方正在输入";
-                break;
-            case TypingMessageContent.TYPING_VOICE:
-                typingDesc = "对方正在录音";
-                break;
-            case TypingMessageContent.TYPING_CAMERA:
-                typingDesc = "对方正在拍照";
-                break;
-            case TypingMessageContent.TYPING_FILE:
-                typingDesc = "对方正在发送文件";
-                break;
-            case TypingMessageContent.TYPING_LOCATION:
-                typingDesc = "对方正在发送位置";
-                break;
-            default:
-                typingDesc = "unknown";
-                break;
-        }
-        setTitle(typingDesc);
-        handler.postDelayed(resetConversationTitleRunnable, 5000);
-    }
-
-    private Runnable resetConversationTitleRunnable = this::resetConversationTitle;
-
-    private void resetConversationTitle() {
-        if (!TextUtils.equals(conversationTitle, getTitle())) {
-            setTitle(conversationTitle);
-            handler.removeCallbacks(resetConversationTitleRunnable);
-        }
-    }
-
-    @Override
-    public void onInputPanelExpanded() {
-        recyclerView.scrollToPosition(adapter.getItemCount() - 1);
-    }
-
-    @Override
-    public void onInputPanelCollapsed() {
-        // do nothing
-    }
-
-    public void toggleMultiMessageMode(UiMessage message) {
-        inputPanel.setVisibility(View.GONE);
-        message.isChecked = true;
-        adapter.setMode(ConversationMessageAdapter.MODE_CHECKABLE);
-        adapter.notifyDataSetChanged();
-        multiMessageActionContainerLinearLayout.setVisibility(View.VISIBLE);
-        setupMultiMessageAction();
-    }
-
-    public void toggleConversationMode() {
-        inputPanel.setVisibility(View.VISIBLE);
-        multiMessageActionContainerLinearLayout.setVisibility(View.GONE);
-        adapter.setMode(ConversationMessageAdapter.MODE_NORMAL);
-        adapter.clearMessageCheckStatus();
-        adapter.notifyDataSetChanged();
-    }
-
-    private void setupMultiMessageAction() {
-        multiMessageActionContainerLinearLayout.removeAllViews();
-        List<MultiMessageAction> actions = MultiMessageActionManager.getInstance().getConversationActions(conversation);
-        for (MultiMessageAction action : actions) {
-
-            action.onBind(this, conversation);
-            TextView textView = new TextView(this);
-            textView.setCompoundDrawablePadding(10);
-            textView.setCompoundDrawablesWithIntrinsicBounds(action.iconResId(), 0, 0, 0);
-            textView.setText(action.title(this));
-            LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT);
-            multiMessageActionContainerLinearLayout.addView(textView, layoutParams);
-
-            textView.setOnClickListener(v -> {
-                List<UiMessage> checkedMessages = adapter.getCheckedMessages();
-                if (action.confirm()) {
-                    new MaterialDialog.Builder(ConversationActivity.this).content(action.confirmPrompt())
-                            .negativeText("取消")
-                            .positiveText("确认")
-                            .onPositive((dialog, which) -> {
-                                action.onClick(checkedMessages);
-                                toggleConversationMode();
-                            })
-                            .build()
-                            .show();
-
-                } else {
-                    action.onClick(checkedMessages);
-                    toggleConversationMode();
-                }
-            });
-        }
-    }
-
-    @Override
-    public void onMessageCheck(UiMessage message, boolean checked) {
-        List<UiMessage> checkedMessages = adapter.getCheckedMessages();
-        setAllClickableChildViewState(multiMessageActionContainerLinearLayout, checkedMessages.size() > 0);
-    }
-
-    private void setAllClickableChildViewState(View view, boolean enable) {
-        if (view instanceof ViewGroup) {
-            for (int i = 0; i < ((ViewGroup) view).getChildCount(); i++) {
-                setAllClickableChildViewState(((ViewGroup) view).getChildAt(i), enable);
-            }
-        }
-        if (view.isClickable()) {
-            view.setEnabled(enable);
-        }
     }
 }
